# adminApi/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{{.Description}}",
        "title": "微服务的管理端api文档demo",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "localhost:9081",
    "basePath": "/api/admin",
    "paths": {
        "/user/edit_pwd": {
            "post": {
                "security": [
                    {
                        "MustToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user   用户操作"
                ],
                "summary": "用户修改密码",
                "parameters": [
                    {
                        "description": "修改密码",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.editPwdReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{code:0,msg:ok}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    },
                    "400": {
                        "description": "{code:1,msg:无效的请求参数}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    },
                    "500": {
                        "description": "{code:-1,msg:服务器故障}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    }
                }
            }
        },
        "/user/info": {
            "get": {
                "security": [
                    {
                        "MustToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user   用户操作"
                ],
                "summary": "获取用户信息及可访问的权限菜单",
                "responses": {
                    "200": {
                        "description": "{code:0,msg:ok}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    },
                    "400": {
                        "description": "{code:1,msg:无效的请求参数}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    },
                    "500": {
                        "description": "{code:-1,msg:服务器故障}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user   用户操作"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{code:0,msg:ok,data:{token:}}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    },
                    "400": {
                        "description": "{code:1,msg:无效的请求参数}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    },
                    "500": {
                        "description": "{code:-1,msg:服务器故障}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "get": {
                "security": [
                    {
                        "MustToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user   用户操作"
                ],
                "summary": "用户登出",
                "responses": {
                    "200": {
                        "description": "{code:0,msg:ok}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    },
                    "400": {
                        "description": "{code:1,msg:无效的请求参数}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    },
                    "500": {
                        "description": "{code:-1,msg:服务器故障}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    }
                }
            }
        },
        "/user_mana/create": {
            "post": {
                "security": [
                    {
                        "MustToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_mana   用户管理模块"
                ],
                "summary": "创建admin用户",
                "parameters": [
                    {
                        "description": "用户资料",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.createAdminUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{code:0,msg:ok}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    },
                    "400": {
                        "description": "{code:1,msg:无效的请求参数}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    },
                    "500": {
                        "description": "{code:-1,msg:服务器故障}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    }
                }
            }
        },
        "/user_mana/delete": {
            "post": {
                "security": [
                    {
                        "MustToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_mana   用户管理模块"
                ],
                "summary": "删除admin用户",
                "parameters": [
                    {
                        "description": "id 列表",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.ids"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{code:0,msg:ok}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    },
                    "400": {
                        "description": "{code:1,msg:无效的请求参数}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    },
                    "500": {
                        "description": "{code:-1,msg:服务器故障}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    }
                }
            }
        },
        "/user_mana/detail": {
            "get": {
                "security": [
                    {
                        "MustToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_mana   用户管理模块"
                ],
                "summary": "获取admin用户详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户uid",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户详情",
                        "schema": {
                            "$ref": "#/definitions/model.AdminUser"
                        }
                    },
                    "400": {
                        "description": "{code:1,msg:无效的请求参数}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    },
                    "500": {
                        "description": "{code:-1,msg:服务器故障}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    }
                }
            }
        },
        "/user_mana/list": {
            "get": {
                "security": [
                    {
                        "MustToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_mana   用户管理模块"
                ],
                "summary": "获取admin用户列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码,默认为1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "返回条数,默认为10",
                        "name": "num",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段,默认为createdAt",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键字",
                        "name": "key",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序规则,默认为DESC",
                        "name": "orderType",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户列表",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.AdminUser"
                            }
                        }
                    },
                    "400": {
                        "description": "{code:1,msg:无效的请求参数}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    },
                    "500": {
                        "description": "{code:-1,msg:服务器故障}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    }
                }
            }
        },
        "/user_mana/setRole": {
            "post": {
                "security": [
                    {
                        "MustToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_mana   用户管理模块"
                ],
                "summary": "为用户添加角色权限",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户uid",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "角色id",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.ids"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{code:0,msg:ok}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    },
                    "400": {
                        "description": "{code:1,msg:无效的请求参数}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    },
                    "500": {
                        "description": "{code:-1,msg:服务器故障}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    }
                }
            }
        },
        "/user_mana/update": {
            "post": {
                "security": [
                    {
                        "MustToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_mana   用户管理模块"
                ],
                "summary": "更新admin用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户uid",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "用户资料",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.updateAdminUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{code:0,msg:ok}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    },
                    "400": {
                        "description": "{code:1,msg:无效的请求参数}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    },
                    "500": {
                        "description": "{code:-1,msg:服务器故障}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    }
                }
            }
        },
        "/user_mana/usersRoleIDList": {
            "get": {
                "security": [
                    {
                        "MustToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_mana   用户管理模块"
                ],
                "summary": "获取用户下的角色ID列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户uid",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{code:0,msg:ok}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    },
                    "400": {
                        "description": "{code:1,msg:无效的请求参数}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    },
                    "500": {
                        "description": "{code:-1,msg:服务器故障}",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponseModel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.ResponseModel": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "handler.createAdminUserReq": {
            "type": "object",
            "required": [
                "password",
                "status",
                "username"
            ],
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号",
                    "type": "string"
                },
                "real_name": {
                    "description": "真实姓名",
                    "type": "string"
                },
                "status": {
                    "description": "状态(1:启用  2.禁用)",
                    "type": "integer"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "handler.editPwdReq": {
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "description": "新密码",
                    "type": "string"
                },
                "old_password": {
                    "description": "旧密码",
                    "type": "string"
                }
            }
        },
        "handler.ids": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "description": "id 列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "handler.login": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "handler.updateAdminUserReq": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号",
                    "type": "string"
                },
                "real_name": {
                    "description": "真实姓名",
                    "type": "string"
                },
                "status": {
                    "description": "状态(1:启用  2.禁用)",
                    "type": "integer"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "model.AdminUser": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号",
                    "type": "string"
                },
                "real_name": {
                    "description": "真实姓名",
                    "type": "string"
                },
                "status": {
                    "description": "状态(1:启用  2.禁用)",
                    "type": "integer"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "MustToken": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}